+++ src/main/java/fr/estecka/variantscit/VCitRegistries.java
@@ -1,6 +1,7 @@
 package fr.estecka.variantscit;
 
 import net.minecraft.util.Identifier;
+import fr.estecka.variantscit.format.INbtInput;
 import fr.estecka.variantscit.format.IStringTransform;
 import fr.estecka.variantscit.format.RegexTransform;
 import fr.estecka.variantscit.format.properties.*;
@@ -9,6 +10,7 @@ public final class VCitRegistries
 {
 	static public final DecodableRegistry<IStringProperty> ITEM_PROPERTIES = new DecodableRegistry<>("property", Identifier.ofVanilla("item_component"), TransformableProperty::CodecOf);
 	static public final DecodableRegistry<IStringTransform> TRANSFORMS = new DecodableRegistry<>("transform", Identifier.ofVanilla("regex"));
+	static public final DecodableRegistry<INbtInput> NBT_INPUTS = new DecodableRegistry<>("type");
 
 	static {
 		ITEM_PROPERTIES.RegisterUnit(Identifier.ofVanilla("axolotl_variant"), AxolotlVariantProperty.UNIT);
@@ -18,13 +20,21 @@ public final class VCitRegistries
 		ITEM_PROPERTIES.RegisterUnit(Identifier.ofVanilla("item_type"), new ItemTypeProperty());
 		ITEM_PROPERTIES.RegisterUnit(Identifier.ofVanilla("painting_variant"), PaintingVariantProperty.UNIT);
 
-		TRANSFORMS.RegisterUnit(Identifier.ofVanilla("noop"), s->s);
-		TRANSFORMS.RegisterUnit(Identifier.ofVanilla("lowercase"), String::toLowerCase);
-		TRANSFORMS.RegisterUnit(Identifier.ofVanilla("discard_path"),      IStringTransform::DiscardPath);
-		TRANSFORMS.RegisterUnit(Identifier.ofVanilla("discard_namespace"), IStringTransform::DiscardNamespace);
+		TRANSFORMS.RegisterUnit(Identifier.ofVanilla("noop"),               s->s);
+		TRANSFORMS.RegisterUnit(Identifier.ofVanilla("lowercase"),          String::toLowerCase);
+		TRANSFORMS.RegisterUnit(Identifier.ofVanilla("discard_path"),       IStringTransform::DiscardPath);
+		TRANSFORMS.RegisterUnit(Identifier.ofVanilla("discard_namespace"),  IStringTransform::DiscardNamespace);
 		TRANSFORMS.RegisterUnit(Identifier.ofVanilla("sanitize"),           IStringTransform.Sanitize("[^a-zA-Z0-9_.-/:]"));
 		TRANSFORMS.RegisterUnit(Identifier.ofVanilla("sanitize_path"),      IStringTransform.Sanitize("[^a-zA-Z0-9_.-/]"));
 		TRANSFORMS.RegisterUnit(Identifier.ofVanilla("sanitize_namespace"), IStringTransform.Sanitize("[^a-zA-Z0-9_.-]"));
-		TRANSFORMS.RegisterMap(Identifier.ofVanilla("regex"), RegexTransform.MAPCODEC);
+		TRANSFORMS.RegisterMap(Identifier.ofVanilla("regex"),               RegexTransform.MAPCODEC);
+
+		NBT_INPUTS.RegisterUnit(Identifier.ofVanilla("auto"),            INbtInput.AUTO);
+		NBT_INPUTS.RegisterUnit(Identifier.ofVanilla("string"),          INbtInput::String);
+		NBT_INPUTS.RegisterUnit(Identifier.ofVanilla("number"),          INbtInput::Number);
+		NBT_INPUTS.RegisterUnit(Identifier.ofVanilla("identifier"),      INbtInput::Identifier);
+		NBT_INPUTS.RegisterUnit(Identifier.ofVanilla("rich_text"),       INbtInput::RichText);
+		NBT_INPUTS.RegisterUnit(Identifier.ofVanilla("rich_text_array"), INbtInput::RichTextArray);
+		NBT_INPUTS.RegisterUnit(Identifier.ofVanilla("primitive"),       INbtInput.Grouped(INbtInput::String, INbtInput::Number));
 	}
 }