+++ src/main/java/fr/estecka/variantscit/format/NbtAdapter.java
@@ -1,34 +1,25 @@
 package fr.estecka.variantscit.format;
 
-import java.util.List;
-import java.util.function.Function;
 import org.jetbrains.annotations.Nullable;
 import com.mojang.serialization.Codec;
 import com.mojang.serialization.MapCodec;
 import com.mojang.serialization.codecs.RecordCodecBuilder;
 import fr.estecka.variantscit.CodecUtil;
-import net.minecraft.nbt.AbstractNbtNumber;
 import net.minecraft.nbt.NbtElement;
-import net.minecraft.nbt.NbtOps;
-import net.minecraft.nbt.NbtString;
-import net.minecraft.text.Text;
-import net.minecraft.text.TextCodecs;
-import net.minecraft.util.Identifier;
-import net.minecraft.util.StringIdentifiable;
 
 public class NbtAdapter
 {
 	static public final MapCodec<NbtAdapter> MAPCODEC = RecordCodecBuilder.mapCodec(builder->builder
 		.group(
 			NbtPath.CODEC.optionalFieldOf("nbtPath", NbtPath.IDENTITY).forGetter(adp -> adp.nbtPath),
-			EInput.CODEC.optionalFieldOf("expect", EInput.PRIMITIVE).forGetter(adp -> adp.type)
+			INbtInput.CODEC.optionalFieldOf("expect", INbtInput.AUTO).forGetter(adp -> adp.type)
 		)
 		.apply(builder, NbtAdapter::new)
 	);
 
 	static public final Codec<NbtAdapter> CODEC = Codec.withAlternative(
 		MAPCODEC.codec(),
-		NbtPath.CODEC.xmap(path->new NbtAdapter(path, EInput.PRIMITIVE), adp->adp.nbtPath)
+		NbtPath.CODEC.xmap(path->new NbtAdapter(path, INbtInput.AUTO), adp->adp.nbtPath)
 	);
 
 	@Deprecated
@@ -36,13 +27,13 @@ public class NbtAdapter
 		NbtPath.CODEC.fieldOf("nbtPath"),
 		NbtPath.DOT_SEPARATED_CODEC.fieldOf("nbtPath"),
 		NbtPath.NBTKEY_CODEC.fieldOf("nbtKey")
-	).xmap((path)->new NbtAdapter(path,EInput.PRIMITIVE), (adp)->adp.nbtPath);
+	).xmap((path)->new NbtAdapter(path,INbtInput.AUTO), (adp)->adp.nbtPath);
+
 
 	private final NbtPath nbtPath;
-	private final EInput type;
+	private final INbtInput type;
 
-
-	protected NbtAdapter(NbtPath nbtPath, EInput type){
+	protected NbtAdapter(NbtPath nbtPath, INbtInput type){
 		this.nbtPath = nbtPath;
 		this.type = type;
 	}
@@ -52,69 +43,10 @@ public class NbtAdapter
 		if (nbt == null)
 			return null;
 
-		String data = this.type.Accept(nbt);
+		String data = this.type.apply(nbt);
 		if (data == null)
 			return null;
 
 		return data;
 	}
-
-	public enum EInput
-	implements StringIdentifiable
-	{
-		PRIMITIVE("primitive", (nbt)->{
-			if (nbt instanceof NbtString)
-				return nbt.asString();
-			else if (nbt instanceof AbstractNbtNumber number)
-				return number.numberValue().toString();
-			else
-				return null;
-		}),
-
-		NUMBER("number", (nbt)-> (nbt instanceof AbstractNbtNumber number) ? number.numberValue().toString() : null),
-		STRING("string", (nbt)-> (nbt instanceof NbtString) ? nbt.asString() : null),
-		IDENTIFIER("identifier", (nbt)-> (nbt instanceof NbtString) ? Identifier.tryParse(nbt.asString()).toString() : null),
-		RICH_TEXT("rich_text", (nbt)->{
-			var text = TextCodecs.STRINGIFIED_CODEC.parse(NbtOps.INSTANCE, nbt);
-			if (text.isSuccess())
-				return text.getOrThrow().getString();
-
-			var multiline = TextCodecs.STRINGIFIED_CODEC.sizeLimitedListOf(256).parse(NbtOps.INSTANCE, nbt);
-			if (multiline.isSuccess())
-				return Concat(multiline.getOrThrow());
-
-			return null;
-		}),
-		;
-
-		static public final Codec<EInput> CODEC = StringIdentifiable.createCodec(EInput::values);
-
-		static private final String Concat(List<Text> lines){
-			StringBuilder builder = new StringBuilder();
-			for (var l : lines) {
-				builder.append(l.getString());
-				builder.append('\n');
-			}
-
-			return builder.toString();
-		}
-
-		private final String name;
-		private final Function<NbtElement,String> lambda;
-
-		private EInput(String name, Function<NbtElement,String> lambda){
-			this.name = name;
-			this.lambda = lambda;
-		}
-
-		public @Nullable String Accept(NbtElement nbt){
-			return this.lambda.apply(nbt);
-		}
-	
-		@Override public String asString(){
-			return this.name;
-		}
-	
-	}
-
 }